Onky try to use gemmiStructure for assembly info if it already exists.

+++ baby-gru/src/components/button-bar/MoorhenMoleculeCardButtonBar.tsx
@@ -32,10 +32,10 @@ type MoorhenMoleculeCardButtonBarPropsType = {
 
 export const MoorhenMoleculeCardButtonBar = (props: MoorhenMoleculeCardButtonBarPropsType) => {
     const dropdownCardButtonRef = useRef<HTMLDivElement>()
-    
+
     const [popoverIsShown, setPopoverIsShown] = useState<boolean>(false)
     const [currentName, setCurrentName] = useState<string>(props.molecule.name);
-    
+
     const dispatch = useDispatch()
     const height = useSelector((state: moorhen.State) => state.sceneSettings.height)
     const makeBackups = useSelector((state: moorhen.State) => state.backupSettings.makeBackups)
@@ -145,30 +145,33 @@ export const MoorhenMoleculeCardButtonBar = (props: MoorhenMoleculeCardButtonBar
         }
     })
 
-    const assemblies = props.molecule.gemmiStructure.assemblies
     let showAssemblies = false
-    for(let i=0; i<assemblies.size(); i++){
-        const assembly = assemblies.get(i)
-        const generators = assembly.generators
-        const n_gen = generators.size()
-        let n_tot_op = 0
-        for (let i_gen=0; i_gen < n_gen; i_gen++) { 
-            const gen = generators.get(i_gen)
-            const operators = gen.operators
-            const n_op = operators.size()
-            n_tot_op += n_op
-            gen.delete()
-            operators.delete()
-        }
-        assembly.delete()
-        generators.delete()
 
-        if(n_tot_op!==60&&n_tot_op!==1){
-            showAssemblies = true
-            break
+    if(props.molecule.gemmiStructure){
+        const assemblies = props.molecule.gemmiStructure.assemblies
+        for(let i=0; i<assemblies.size(); i++){
+            const assembly = assemblies.get(i)
+            const generators = assembly.generators
+            const n_gen = generators.size()
+            let n_tot_op = 0
+            for (let i_gen=0; i_gen < n_gen; i_gen++) {
+                const gen = generators.get(i_gen)
+                const operators = gen.operators
+                const n_op = operators.size()
+                n_tot_op += n_op
+                gen.delete()
+                operators.delete()
+            }
+            assembly.delete()
+            generators.delete()
+
+            if(n_tot_op!==60&&n_tot_op!==1){
+                showAssemblies = true
+                break
+            }
         }
+        assemblies.delete()
     }
-    assemblies.delete()
 
     if(showAssemblies){
         Object.keys(bioMolButtons).forEach(key => {
@@ -186,10 +189,10 @@ export const MoorhenMoleculeCardButtonBar = (props: MoorhenMoleculeCardButtonBar
     }
 
     compressedButtons.push(
-        <MoorhenDeleteDisplayObjectMenuItem 
+        <MoorhenDeleteDisplayObjectMenuItem
             key="deleteDisplayObjectMenuItem"
-            setPopoverIsShown={setPopoverIsShown} 
-            glRef={props.glRef} 
+            setPopoverIsShown={setPopoverIsShown}
+            glRef={props.glRef}
             item={props.molecule} />
     )
